refactor: simplify test setup
This clears the codebase for better readability 
and maintains focus on relevant tests for the 
BraintrustTracer and associated components
